moved Resampler in its own file and mark pub(crate)
