+++ src/buffer.rs
@@ -1,7 +1,6 @@
 //! General purpose audio signal data structures
 use std::sync::Arc;
 
-use crate::media::MediaStream;
 use crate::render::AudioRenderQuantum;
 use crate::SampleRate;
 
@@ -28,8 +27,6 @@ pub struct AudioBuffer {
     channels: Vec<ChannelData>,
     sample_rate: SampleRate,
 }
-
-use std::error::Error;
 
 impl AudioBuffer {
     /// Allocate a silent audiobuffer with [`AudioBufferOptions`]
@@ -321,112 +318,6 @@ impl ChannelData {
     }
 }
 
-/// Sample rate converter and buffer chunk splitter.
-///
-/// A `MediaElement` can be wrapped inside a `Resampler` to yield AudioBuffers of the desired sample_rate and length
-///
-/// ```
-/// use web_audio_api::SampleRate;
-/// use web_audio_api::buffer::{AudioBuffer, Resampler};
-///
-/// // construct an input of 3 chunks of 5 samples
-/// let samples = vec![vec![1., 2., 3., 4., 5.]];
-/// let input_buf = AudioBuffer::from(samples, SampleRate(44_100));
-/// let input = vec![input_buf; 3].into_iter().map(|b| Ok(b));
-///
-/// // resample to chunks of 10 samples
-/// let mut resampler = Resampler::new(SampleRate(44_100), 10, input);
-///
-/// // first chunk contains 10 samples
-/// let next = resampler.next().unwrap().unwrap();
-/// assert_eq!(next.length(), 10);
-/// assert_eq!(next.get_channel_data(0)[..], vec![
-///     1., 2., 3., 4., 5.,
-///     1., 2., 3., 4., 5.,
-/// ][..]);
-///
-/// // second chunk contains 5 samples of signal, and 5 silent
-/// let next = resampler.next().unwrap().unwrap();
-/// assert_eq!(next.length(), 10);
-/// assert_eq!(next.get_channel_data(0)[..], vec![
-///     1., 2., 3., 4., 5.,
-///     0., 0., 0., 0., 0.,
-/// ][..]);
-///
-/// // no further chunks
-/// assert!(resampler.next().is_none());
-/// ```
-pub struct Resampler<I> {
-    /// desired sample rate
-    sample_rate: SampleRate,
-    /// desired sample length
-    sample_len: usize,
-    /// input stream
-    input: I,
-    /// internal buffer
-    buffer: Option<AudioBuffer>,
-}
-
-impl<M: MediaStream> Resampler<M> {
-    pub fn new(sample_rate: SampleRate, sample_len: usize, input: M) -> Self {
-        Self {
-            sample_rate,
-            sample_len,
-            input,
-            buffer: None,
-        }
-    }
-}
-
-impl<M: MediaStream> Iterator for Resampler<M> {
-    type Item = Result<AudioBuffer, Box<dyn Error + Send + Sync>>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let mut buffer = match self.buffer.take() {
-            None => match self.input.next() {
-                None => return None,
-                Some(Err(e)) => return Some(Err(e)),
-                Some(Ok(mut data)) => {
-                    data.resample(self.sample_rate);
-                    data
-                }
-            },
-            Some(data) => data,
-        };
-
-        while buffer.length() < self.sample_len {
-            // buffer is smaller than desired len
-            match self.input.next() {
-                None => {
-                    let options = AudioBufferOptions {
-                        number_of_channels: buffer.number_of_channels(),
-                        length: self.sample_len - buffer.length(),
-                        sample_rate: self.sample_rate,
-                    };
-
-                    let padding = AudioBuffer::new(options);
-                    buffer.extend(&padding);
-
-                    return Some(Ok(buffer));
-                }
-                Some(Err(e)) => return Some(Err(e)),
-                Some(Ok(mut data)) => {
-                    data.resample(self.sample_rate);
-                    buffer.extend(&data)
-                }
-            }
-        }
-
-        if buffer.length() == self.sample_len {
-            return Some(Ok(buffer));
-        }
-
-        self.buffer = Some(buffer.split_off(self.sample_len));
-
-        Some(Ok(buffer))
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use float_eq::assert_float_eq;
@@ -658,60 +549,5 @@ mod tests {
             abs_all <= 0.
         );
         assert_eq!(buffer.sample_rate_raw().0, 100);
-    }
-
-    #[test]
-    fn test_resampler_concat() {
-        let channel = ChannelData::from(vec![1., 2., 3., 4., 5.]);
-        let input_buf = AudioBuffer::from_channels(vec![channel], SampleRate(44_100));
-        let input = vec![input_buf; 3].into_iter().map(Ok);
-        let mut resampler = Resampler::new(SampleRate(44_100), 10, input);
-
-        let next = resampler.next().unwrap().unwrap();
-        assert_eq!(next.length(), 10);
-        assert_float_eq!(
-            next.channel_data(0).as_slice(),
-            &[1., 2., 3., 4., 5., 1., 2., 3., 4., 5.,][..],
-            abs_all <= 0.
-        );
-
-        let next = resampler.next().unwrap().unwrap();
-        assert_eq!(next.length(), 10);
-        assert_float_eq!(
-            next.channel_data(0).as_slice(),
-            &[1., 2., 3., 4., 5., 0., 0., 0., 0., 0.][..],
-            abs_all <= 0.
-        );
-
-        assert!(resampler.next().is_none());
-    }
-
-    #[test]
-    fn test_resampler_split() {
-        let channel = ChannelData::from(vec![1., 2., 3., 4., 5., 6., 7., 8., 9., 10.]);
-        let input_buf = Ok(AudioBuffer::from_channels(
-            vec![channel],
-            SampleRate(44_100),
-        ));
-        let input = vec![input_buf].into_iter();
-        let mut resampler = Resampler::new(SampleRate(44_100), 5, input);
-
-        let next = resampler.next().unwrap().unwrap();
-        assert_eq!(next.length(), 5);
-        assert_float_eq!(
-            next.channel_data(0).as_slice(),
-            &[1., 2., 3., 4., 5.][..],
-            abs_all <= 0.
-        );
-
-        let next = resampler.next().unwrap().unwrap();
-        assert_eq!(next.length(), 5);
-        assert_float_eq!(
-            next.channel_data(0).as_slice(),
-            &[6., 7., 8., 9., 10.][..],
-            abs_all <= 0.
-        );
-
-        assert!(resampler.next().is_none());
     }
 }
