+++ src/node/media_element.rs
@@ -1,7 +1,7 @@
-use crate::buffer::Resampler;
 use crate::context::{AudioContextRegistration, BaseAudioContext};
 use crate::control::{Controller, Scheduler};
 use crate::media::MediaElement;
+use crate::media::Resampler;
 use crate::RENDER_QUANTUM_SIZE;
 
 use super::{
